pyproject.toml
[project]
name = "mcpred"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "click>=8.2.1",
    "mcp>=1.12.3",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
    "websockets>=15.0.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
]
